srcs/Two_Factor_authentication/pyproject.toml
[tool.poetry]
name = "42-ft-transcendence"
version = "0.1.0"
description = ""
authors = ["Jhonata <jhsocial@icloud.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django = "^5.1.2"
dj-database-url = "^2.2.0"
psycopg2-binary = "2.9.6"
python-decouple = "^3.8"
whitenoise = "^6.7.0"
pillow = "11.0.0" # work with jpg images (for the avatars)
djangorestframework-simplejwt= "*" #token generation
pyaml = "14.04.3" #render the OpenAPI view
uritemplate = "4.1.1" #render the OpenAPI view
drf-spectacular = { version = "0.27.2", extras = ["swagger-ui"] } # provide more control on OpenAPI views with Swagger UI
django-cors-headers = "^4.6.0"
qrcode = "^8.0"
pyotp = "^2.9.0"
jwt = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
